refactor(project): break into classes, switch to Axios
