+++ .gitignore
@@ -1,7 +1,6 @@
 .idea
 .vscode
 _scratch
-lib
 dist
 docs
 .DS_Store
