+++ package.json
@@ -7,31 +7,28 @@
     },
     "dependencies": {
         "@expo/plist": "0.0.20",
+        "axios": "^1.6.7",
         "form-data": "4.0.0",
-        "got": "11.8.6",
         "uuid": "8.3.2"
     },
     "description": "",
     "devDependencies": {
-        "@pliancy/eslint-config-ts": "1.0.0",
+        "@pliancy/eslint-config-ts": "1.1.0",
         "@pliancy/semantic-release-config-npm": "2.2.0",
-        "@types/got": "9.6.12",
         "@types/jest": "29.5.5",
-        "@types/node": "15.14.9",
-        "@types/uuid": "8.3.4",
+        "@types/node": "20.11.20",
+        "@types/uuid": "9.0.8",
         "commitizen": "4.3.0",
         "cpy-cli": "5.0.0",
         "cz-conventional-changelog": "3.3.0",
-        "gh-pages": "3.2.3",
         "husky": "8.0.3",
         "jest": "29.7.0",
         "npm-run-all": "4.1.5",
-        "open": "9.1.0",
-        "pinst": "^3.0.0",
-        "rimraf": "3.0.2",
+        "open": "10.0.4",
+        "pinst": "3.0.0",
+        "rimraf": "5.0.5",
         "ts-jest": "29.1.1",
-        "typedoc": "0.25.1",
-        "typescript": "4.9.5"
+        "typescript": "5.3.3"
     },
     "keywords": [
         "addigy",
