--- src/index.spec.ts
@@ -1,3 +1,0 @@
-describe('Addigy', () => {
-    it('is a placeholder', () => expect(true).toBe(true))
-})
