+++ src/index.ts
@@ -1,1578 +1,19 @@
-import got from 'got'
-import { v4 as uuidv4 } from 'uuid'
-import {
-    CreateSoftware,
-    CreateWebContentFilterPayload,
-    CustomFact,
-    CustomProfilePayload,
-    Extension,
-    FilevaultPayload,
-    FilevaultRequest,
-    IAddigyConfig,
-    IAddigyInternalAuthObject,
-    KernalExtensionPayload,
-    MdmConfigurationPayload,
-    NotificationSettings,
-    NotificationSettingsPayload,
-    PPPCInput,
-    PPPCPayload,
-    PPPCService,
-    ServiceManagementPayload,
-    ServiceManagementPayloadRule,
-    Software,
-    SupportedOsVersions,
-    SystemExtensionPayload,
-    WebContentFilterPayload,
-} from './types'
-import plist from '@expo/plist'
-
-export * from './types'
-
-enum AlertStatus {
-    Acknowledged = 'Acknowledged',
-    Resolved = 'Resolved',
-    Unattended = 'Unattended',
-}
-
-enum UserRoles {
-    Owner = 'power',
-    Admin = 'admin',
-    User = 'user',
-}
-
-export class Addigy {
-    config: IAddigyConfig
-    domain: string
-    reqHeaders: any
-
-    constructor(_config: IAddigyConfig) {
-        this.config = _config
-        this.reqHeaders = {
-            'content-type': 'application/json',
-            accept: 'application/json',
-            'client-id': this.config.clientId,
-            'client-secret': this.config.clientSecret,
-        }
-        this.domain = 'https://prod.addigy.com/api'
-    }
-
-    //
-    // Instructions
-    //
-
-    async getPolicyInstructions(
-        policyId: string,
-        provider: string = 'ansible-profile',
-    ): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/policies/instructions?provider=${provider}&policy_id=${policyId}`,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createPolicyInstructions(policyId: string, instructionId: string): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(`${this.domain}/policies/instructions`, {
-                headers: this.reqHeaders,
-                method: 'POST',
-                body: JSON.stringify({
-                    instruction_id: instructionId,
-                    policy_id: policyId,
-                }),
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async deletePolicyInstructions(
-        policyId: string,
-        instructionId: string,
-        provider: string = 'ansible-profile',
-    ): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/policies/instructions?policy_id=${policyId}&instruction_id=${instructionId}&provider=${provider}`,
-                {
-                    headers: this.reqHeaders,
-                    method: 'DELETE',
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Devices
-    //
-
-    async getOnlineDevices(): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(`${this.domain}/devices/online`, {
-                headers: this.reqHeaders,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getDevices(): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(`${this.domain}/devices`, {
-                headers: this.reqHeaders,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getPolicyDevices(policyId: string): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/policies/devices?policy_id=${policyId}`,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async updateDevicePolicy(policyId: string, agentId: string): Promise<object[]> {
-        let postBody: any = {
-            policy_id: policyId,
-            agent_id: agentId,
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/policies/devices`, {
-                headers: {
-                    'client-id': this.config.clientId,
-                    'client-secret': this.config.clientSecret,
-                },
-                method: 'POST',
-                form: true,
-                body: postBody,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Alerts
-    //
-
-    async getAlerts(
-        status: AlertStatus,
-        page: number = 1,
-        pageLength: number = 10,
-    ): Promise<object[]> {
-        let statusUri = ''
-        if (status) {
-            statusUri = `&status=${status}`
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/alerts?page=${page}&per_page=${pageLength}` + statusUri,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Policies
-    //
-
-    async getPolicies(): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(`${this.domain}/policies`, {
-                headers: this.reqHeaders,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getPolicyDetails(
-        policyId: string,
-        provider: string = 'ansible-profile',
-    ): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/policies/details?provider=${provider}&policy_id=${policyId}`,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createPolicy(
-        name: string,
-        parentId?: string,
-        icon?: string,
-        color?: string,
-    ): Promise<object[]> {
-        let postBody: any = {
-            name: name,
-        }
-
-        if (icon !== undefined) {
-            postBody['icon'] = icon
-        }
-
-        if (color !== undefined) {
-            postBody['color'] = color
-        }
-
-        if (parentId !== undefined) {
-            postBody['parent_id'] = parentId
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/policies`, {
-                method: 'POST',
-                form: true,
-                body: postBody,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Maintenance
-    //
-
-    async getMaintenance(page: number = 1, pageLenth: number = 10): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/maintenance?page=${page}&per_page=${pageLenth}`,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Applications
-    //
-
-    async getInstalledApplications(): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(`${this.domain}/applications`, {
-                headers: this.reqHeaders,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Profiles
-    //
-
-    async getProfiles(instructionId?: string): Promise<object[]> {
-        let instructionUri = ''
-        if (instructionId !== undefined) {
-            instructionUri = `?instruction_id=${instructionId}`
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/profiles` + instructionUri, {
-                headers: this.reqHeaders,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createProfile(name: string, payloads: object[]): Promise<object[]> {
-        let postBody: any = {
-            name: name,
-            payloads: payloads,
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/profiles`, {
-                headers: this.reqHeaders,
-                method: 'POST',
-                json: postBody,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async updateProfile(instructionId: string, payloads: object[]): Promise<object[]> {
-        let postBody: any = {
-            instruction_id: instructionId,
-            payloads: payloads,
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/profiles`, {
-                headers: this.reqHeaders,
-                method: 'PUT',
-                json: postBody,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async deleteProfile(instructionId: string): Promise<object[]> {
-        let postBody: any = {
-            instruction_id: instructionId,
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/profiles`, {
-                headers: this.reqHeaders,
-                method: 'DELETE',
-                json: postBody,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Commands
-    //
-
-    async runCommand(agentIds: string[], command: string): Promise<object[]> {
-        let postBody: any = {
-            agent_ids: agentIds,
-            command: command,
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/devices/commands`, {
-                headers: this.reqHeaders,
-                method: 'POST',
-                json: postBody,
-            })
-            return res.body
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getCommandOutput(actionId: string, agentId: string): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/devices/output?action_id=${actionId}&agentid=${agentId}`,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Public Software
-    //
-
-    async getPublicSoftware(): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(`${this.domain}/catalog/public`, {
-                headers: this.reqHeaders,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // Custom Software
-    //
-
-    async getCustomSoftware(): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(`${this.domain}/custom-software`, {
-                headers: this.reqHeaders,
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getCustomSoftwareAllVersions(softwareId: string): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/custom-software?identifier=${softwareId}`,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getCustomSoftwareSpecificVersion(instructionId: string): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                `${this.domain}/custom-software?instructionid=${instructionId}`,
-                { headers: this.reqHeaders },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getFileUploadUrl(fileName: string, contentType?: string): Promise<string> {
-        const headers = {
-            'client-Id': this.config.clientId,
-            'client-Secret': this.config.clientSecret,
-            'file-name': fileName,
-            'content-type': contentType ?? 'application/octet-stream',
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                `https://file-manager-prod.addigy.com/api/upload/url`,
-                {
-                    headers: headers,
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async uploadFile(uploadUrl: string, file: object, contentType?: string): Promise<object[]> {
-        const headers = {
-            'content-type': contentType ?? 'application/octet-stream',
-        }
-
-        try {
-            let res = await this._addigyRequest(`${uploadUrl}`, {
-                headers: headers,
-                body: file,
-                method: 'PUT',
-            })
-            return res.body
-        } catch (err) {
-            throw err
-        }
-    }
-
-    /**
-     * V1 API - Creates a custom software object in Addigy.
-     * @param baseIdentifier
-     * @param version
-     * @param downloads
-     * @param installationScript
-     * @param condition
-     * @param removeScript
-     * @param priority
-     * @returns
-     */
-    async createCustomSoftware(
-        baseIdentifier: string,
-        version: string,
-        downloads: string[],
-        installationScript: string,
-        condition: string,
-        removeScript: string,
-        priority = 10,
-    ): Promise<object[]> {
-        let postBody: any = {
-            base_identifier: baseIdentifier,
-            version,
-            downloads,
-            installation_script: installationScript,
-            condition,
-            remove_script: removeScript,
-            priority,
-        }
-
-        try {
-            let res = await this._addigyRequest(`${this.domain}/custom-software`, {
-                headers: this.reqHeaders,
-                method: 'POST',
-                json: postBody,
-            })
-            // Fun fact! This endpoint returns an empty string when successful. Yes, that is correct, an empty string...
-            return res.body
-        } catch (err) {
-            throw err
-        }
-    }
-
-    //
-    // The following endpoints use Addigy's internal API. Use at your own risk.
-    //
-
-    /**
-     *  Internal API - Creates a custom software object in Addigy. This is the internal API, so it is subject to change.
-     *  Different from the V1 API, this allows for things like setting priority.
-     * @param authObject
-     * @param software
-     * @returns
-     */
-    async createSoftwareInternal(
-        authObject: IAddigyInternalAuthObject,
-        software: CreateSoftware,
-    ): Promise<Software> {
-        const res = await this._addigyRequest('https://app.addigy.com/api/software', {
-            headers: {
-                Cookie: `auth_token=${authObject.authToken};`,
-                origin: 'https://app-prod.addigy.com',
-            },
-            method: 'POST',
-            json: software,
-        })
-        return JSON.parse(res.body)
-    }
-
-    async getUsers(authObject: IAddigyInternalAuthObject): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest('https://app-prod.addigy.com/api/account', {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                },
-                method: 'GET',
-            })
-            return JSON.parse(res.body).users
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createUser(
-        authObject: IAddigyInternalAuthObject,
-        email: string,
-        name: string,
-        policies: string[] = [],
-        role: UserRoles | string,
-        phone?: string,
-    ): Promise<object[]> {
-        let postBody: any = {
-            name: name,
-            email: email,
-            policies: policies,
-            role: role,
-        }
-
-        if (phone !== undefined) {
-            postBody['phone'] = phone
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/cloud/users/user',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: postBody,
-                },
-            )
-            return res.body
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async updateUser(
-        authObject: IAddigyInternalAuthObject,
-        email: string,
-        name: string,
-        policies: string[] = [],
-        role: string,
-        phone?: string,
-    ): Promise<object[]> {
-        let postBody: any = {
-            id: '',
-            uid: '', // this has to be blank on th PUT for some reason
-            name: name,
-            authanvil_tfa_username: '',
-            email: email,
-            phone: '',
-            role: role,
-            addigy_role: '', // this also has to be blank
-            policies: policies,
-        }
-
-        if (phone !== undefined) {
-            postBody['phone'] = phone
-        }
-
-        try {
-            // find userId that corresponds to the provided email
-            let users: Array<any> = await this.getUsers(authObject)
-            let user: any = users.find((element) => element.email === email)
-            if (!user) throw new Error(`No user with email ${email} exists.`)
-
-            postBody['id'] = user.id // Addigy requires the user ID to be both in the post body and in the REST URI
-
-            let res = await this._addigyRequest(
-                `https://app-prod.addigy.com/api/cloud/users/user/${
-                    user.id
-                }?user_email=${encodeURIComponent(user.email)}`,
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'PUT',
-                    json: postBody,
-                },
-            )
-            return res.body // returns `ok` if successful...
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async deleteUser(authObject: IAddigyInternalAuthObject, email: string): Promise<object[]> {
-        try {
-            // find userId that corresponds to the provided email
-            let users: Array<any> = await this.getUsers(authObject)
-            let user: any = users.find((element) => element.email === email)
-            if (!user) throw new Error(`No user with email ${email} exists.`)
-
-            let res = await this._addigyRequest(
-                `https://app-prod.addigy.com/api/cloud/users/user/${
-                    user.id
-                }?user_email=${encodeURIComponent(email)}`,
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'DELETE',
-                },
-            )
-
-            return JSON.parse(res.body) // this will return "ok" if successful.
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getBillingData(authObject: IAddigyInternalAuthObject): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/billing/get_chargeover_billing_data',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                        email: authObject.emailAddress,
-                        orgid: authObject.orgId,
-                    },
-                    method: 'GET',
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getApiIntegrations(authObject: IAddigyInternalAuthObject): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest('https://prod.addigy.com/accounts/api/keys/get/', {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                    email: authObject.emailAddress,
-                    orgid: authObject.orgId,
-                },
-                method: 'GET',
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createApiIntegration(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-    ): Promise<object> {
-        let postBody: any = {
-            name,
-        }
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/integrations/keys',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: postBody,
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async deleteApiIntegration(
-        authObject: IAddigyInternalAuthObject,
-        objectId: string,
-    ): Promise<object> {
-        try {
-            let res = await this._addigyRequest(
-                `https://app-prod.addigy.com/api/integrations/keys?id=${objectId}`,
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'DELETE',
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getScreenconnectLinks(
-        authObject: IAddigyInternalAuthObject,
-        sessionId: string,
-        agentId?: string,
-    ): Promise<object[]> {
-        // in most (all?) cases tested, the agentId and sessionId are identical, but they are independently passed in the API call
-        agentId = agentId ? agentId : sessionId
-
-        let postBody = {
-            sessionId: sessionId,
-            agentid: agentId,
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/devices/screenconnect/links',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                        email: authObject.emailAddress,
-                        orgid: authObject.orgId,
-                    },
-                    method: 'POST',
-                    json: postBody,
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    /**
-     *
-     * @param name
-     * Name of the Policy
-     * @param allowOverrides
-     * Users can approve additional kernel extensions not explicitly allowed by configuration profiles.
-     * @param kernalExtensions
-     * An Object to pass through Team Identifiers and Kernal Extensions
-     * allowedTeamIdentifiers: List of Team Identifiers that define which validly signed kernel extensions will be allowed to load
-     * allowedKernalExtensions: Bundle identifer and team identifer of kernel extension that will be allowed. Use An empty team identifier for unsigned legacy kernel extensions
-     */
-
-    async createKernelExtensionPolicy(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        allowOverrides: boolean,
-        kernalExtensions: {
-            allowedTeamIdentifiers?: string[]
-            allowedKernalExtensions?: Extension[]
-        },
-    ): Promise<any> {
-        let payloadUUID = uuidv4()
-        let groupUUID = uuidv4()
-
-        const payload: KernalExtensionPayload = {
-            addigy_payload_type:
-                'com.addigy.syspolicy.kernel-extension-policy.com.apple.syspolicy.kernel-extension-policy',
-            payload_type: 'com.apple.syspolicy.kernel-extension-policy',
-            payload_version: 1,
-            payload_identifier: `com.addigy.syspolicy.kernel-extension-policy.com.apple.syspolicy.kernel-extension-policy.${groupUUID}`,
-            payload_uuid: payloadUUID,
-            payload_group_id: groupUUID,
-            payload_enabled: true,
-            payload_display_name: name,
-            allow_user_overrides: allowOverrides,
-            allowed_kernel_extensions: {},
-            allowed_team_identifiers: [],
-        }
-        if (kernalExtensions.allowedKernalExtensions?.length) {
-            kernalExtensions.allowedKernalExtensions.forEach((e) => {
-                payload.allowed_kernel_extensions[e.teamIdentifier] = e.bundleIdentifiers
-            })
-        }
-        if (kernalExtensions.allowedTeamIdentifiers?.length) {
-            kernalExtensions.allowedTeamIdentifiers.forEach((e) => {
-                payload.allowed_team_identifiers.push(e)
-            })
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: { payloads: [payload] },
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    /**
-     *
-     * @param name
-     * Name of the Policy
-     * @param allowOverrides
-     * Users can approve additional kernel extensions not explicitly allowed by configuration profiles.
-     * @param systemExtensions
-     * An Object to pass through Allowed System Extensions, Allowed System Extension Types and Allowed Team Identifiers
-     * allowedSystemExtensions: A dictionary of system extensions that will always be approved on the machine. The dictionary maps the team identifiers (keys) to arrays of bundle identifiers, where the bundle identifier defines the system extension to be installed.
-     * allowedSystemExtensionTypes: A dictionary mapping a team identifier to an array of strings, where each string is a type of system extension that may be installed for that team identifier
-     * allowedTeamIdentifiers: An array, of team identifiers, that defines valid, signed system extensions which are allowed to load. All system extensions signed with any of the specified team identifiers will be considered to be approved.
-     */
-
-    async createSystemExtensionPolicy(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        allowOverrides: boolean,
-        systemExtensions: {
-            allowedSystemExtensions?: Extension[]
-            allowedSystemExtensionTypes?: Extension[]
-            allowedTeamIdentifiers?: string[]
-        },
-    ): Promise<any> {
-        const groupUUID = uuidv4()
-
-        const payload: SystemExtensionPayload = {
-            addigy_payload_type:
-                'com.addigy.syspolicy.system-extension-policy.com.apple.system-extension-policy',
-            payload_type: 'com.apple.system-extension-policy',
-            payload_version: 1,
-            payload_identifier: `com.addigy.syspolicy.system-extension-policy.com.apple.system-extension-policy.${groupUUID}`,
-            payload_uuid: uuidv4(),
-            payload_group_id: groupUUID,
-            payload_enabled: true,
-            payload_display_name: name,
-            allowed_system_extensions: {},
-            allowed_system_extensions_types: {},
-            allowed_team_identifiers: [],
-            allow_user_overrides: allowOverrides,
-        }
-        if (systemExtensions.allowedSystemExtensions?.length) {
-            systemExtensions.allowedSystemExtensions.forEach((e) => {
-                payload.allowed_system_extensions[e.teamIdentifier] = e.bundleIdentifiers
-            })
-        }
-        if (systemExtensions.allowedSystemExtensionTypes?.length) {
-            systemExtensions.allowedSystemExtensionTypes.forEach((e) => {
-                payload.allowed_system_extensions_types[e.teamIdentifier] = e.bundleIdentifiers
-            })
-        }
-        if (systemExtensions.allowedTeamIdentifiers?.length) {
-            systemExtensions.allowedTeamIdentifiers.forEach((e) => {
-                payload.allowed_team_identifiers.push(e)
-            })
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: { payloads: [payload] },
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createNotificationSettingsPolicy(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        notificationSettings: NotificationSettings[],
-    ): Promise<any> {
-        const groupUUID = uuidv4()
-
-        const payload: NotificationSettingsPayload = {
-            addigy_payload_type: 'com.addigy.notifications.com.apple.notificationsettings',
-            payload_type: 'com.apple.notificationsettings',
-            payload_version: 1,
-            payload_identifier: `com.addigy.notifications.com.apple.notificationsettings.${groupUUID}`,
-            payload_uuid: uuidv4(),
-            payload_group_id: groupUUID,
-            payload_display_name: name,
-            notification_settings: notificationSettings,
-        }
-
-        let res = await this._addigyRequest(
-            'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-            {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                },
-                method: 'POST',
-                json: { payloads: [payload] },
-            },
-        )
-        return JSON.parse(res.body)
-    }
-
-    async createCustomProfile(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        customProfileText: string,
-        supportedOsVersions: SupportedOsVersions,
-        payloadScope: 'System' | 'User' = 'System',
-        is_profile_signed = false,
-    ): Promise<any> {
-        const groupUUID = uuidv4()
-
-        const customProfileJson = plist.parse(customProfileText)
-        // Keys for customProfileJson need to be snake_case
-        const updateCustomProfileJson = Object.entries(customProfileJson).reduce(
-            (acc, [key, value]) => {
-                acc[this.toSnakeCase(key)] = value
-                return acc
-            },
-            {} as any,
-        )
-        const customProfileBase64 = Buffer.from(customProfileText).toString('base64')
-
-        const payload: CustomProfilePayload = {
-            addigy_payload_type: 'com.addigy.custom.mdm.payload',
-            payload_type: 'custom',
-            payload_version: 1,
-            payload_identifier: `com.addigy.custom.mdm.payload.${groupUUID}`,
-            payload_uuid: `custom-profile-${uuidv4()}`,
-            payload_group_id: groupUUID,
-            payload_display_name: name,
-            is_profile_signed,
-            profile_json_data: updateCustomProfileJson,
-            decoded_profile_content: customProfileText,
-            custom_profile_content: customProfileBase64,
-            supported_os_versions: supportedOsVersions,
-            payload_scope: payloadScope,
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: { payloads: [payload] },
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createMdmProfile(authObject: IAddigyInternalAuthObject, mdmProfile: any) {
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: { payloads: mdmProfile },
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async createServiceManagementPolicy(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        rules: ServiceManagementPayloadRule[],
-        priority = 9,
-    ): Promise<any> {
-        const groupUUID = uuidv4()
-
-        const payload: ServiceManagementPayload = {
-            addigy_payload_type: 'com.addigy.servicemanagement.com.apple.servicemanagement',
-            addigy_payload_version: 0,
-            has_manifest: false,
-            payload_display_name: name,
-            payload_enabled: false,
-            payload_group_id: groupUUID,
-            payload_identifier: `com.addigy.servicemanagement.com.apple.servicemanagement.${groupUUID}`,
-            payload_priority: priority,
-            payload_type: 'com.apple.servicemanagement',
-            payload_uuid: uuidv4(),
-            payload_version: 1,
-            policy_restricted: false,
-            requires_device_supervision: false,
-            requires_mdm_profile_approved: false,
-            supported_os_versions: null,
-            rules,
-        }
-
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: { payloads: [payload] },
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    /**
-         @param {string} payloadName - Name of the profile
-         @param {string} userDefinedName - Name of the filter to be displayed in the User
-        @param {string} pluginBundleId - Bundle ID of the plugin to be used for filtering
-    */
-    async createWebContentFilterPolicy(
-        authObject: IAddigyInternalAuthObject,
-        payloadName: string,
-        webContentPayload: CreateWebContentFilterPayload,
-        priority = 9,
-    ) {
-        const groupUUID = uuidv4()
-        const payload: WebContentFilterPayload = {
-            addigy_payload_type: 'com.addigy.webcontent-filter.com.apple.webcontent-filter',
-            addigy_payload_version: 2,
-            auto_filter_enabled: null,
-            blacklisted_urls: null,
-            content_filter_uuid: null,
-            filter_browsers: null,
-            filter_data_provider_bundle_identifier: null,
-            filter_data_provider_designated_requirement: null,
-            filter_packet_provider_bundle_identifier: null,
-            filter_packet_provider_designated_requirement: null,
-            filter_packets: null,
-            filter_sockets: true,
-            filter_type: 'Plugin',
-            has_manifest: false,
-            organization: null,
-            password: null,
-            payload_display_name: payloadName,
-            payload_enabled: true,
-            payload_group_id: groupUUID,
-            payload_identifier: `com.addigy.webcontent-filter.com.apple.webcontent-filter.${groupUUID}`,
-            payload_priority: priority,
-            payload_type: 'com.apple.webcontent-filter',
-            payload_uuid: uuidv4(),
-            payload_version: 1,
-            permitted_urls: null,
-            policy_restricted: false,
-            requires_device_supervision: false,
-            requires_mdm_profile_approved: false,
-            server_address: null,
-            supported_os_versions: null,
-            user_name: null,
-            vendor_config: null,
-            white_listed_bookmarks: null,
-            ...webContentPayload,
-        }
-
-        let res = await this._addigyRequest(
-            'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-            {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                },
-                method: 'POST',
-                json: { payloads: [payload] },
-            },
-        )
-        return JSON.parse(res.body)
-    }
-
-    async createFilevaultPolicy(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        filevault: FilevaultRequest,
-        payloadPriority: number = 1,
-    ) {
-        const groupUUID = uuidv4()
-        const encryptCertPayloadUUID = uuidv4()
-
-        const basePayload = {
-            payload_display_name: name,
-            payload_version: 1,
-            payload_group_id: groupUUID,
-            addigy_payload_version: 0,
-            payload_priority: payloadPriority,
-        }
-
-        const payloads: FilevaultPayload[] = [
-            {
-                ...basePayload,
-                payload_type: 'com.apple.MCX.FileVault2',
-                addigy_payload_type: 'com.addigy.securityAndPrivacy.com.apple.MCX.FileVault2',
-                payload_identifier: `com.addigy.securityAndPrivacy.com.apple.MCX.FileVault2.${groupUUID}`,
-                payload_uuid: uuidv4(),
-                enable: filevault.enable ? 'On' : 'Off',
-                defer: filevault.defer,
-                use_recovery_key: true,
-                show_recovery_key:
-                    filevault.showRecoveryKey === undefined ? null : filevault.showRecoveryKey,
-                defer_dont_ask_at_user_logout:
-                    filevault.deferDontAskAtUserLogout === undefined
-                        ? null
-                        : filevault.deferDontAskAtUserLogout,
-                defer_force_at_user_login_max_bypass_attempts:
-                    filevault.deferForceAtUserLoginMaxBypassAttempts === undefined
-                        ? null
-                        : filevault.deferForceAtUserLoginMaxBypassAttempts,
-            },
-            {
-                ...basePayload,
-                payload_type: 'com.apple.MCX',
-                addigy_payload_type: 'com.addigy.securityAndPrivacy.com.apple.MCX',
-                payload_identifier: `com.addigy.securityAndPrivacy.com.apple.MCX.${groupUUID} `,
-                payload_uuid: uuidv4(),
-                destroy_fv_key_on_standby:
-                    filevault.destroyFvKeyOnStandby === undefined
-                        ? null
-                        : filevault.destroyFvKeyOnStandby,
-                dont_allow_fde_disable: true,
-            },
-        ]
-
-        if (filevault.escrowRecoveryKey)
-            payloads.push(
-                {
-                    ...basePayload,
-                    addigy_payload_type: 'com.addigy.securityAndPrivacy.com.apple.security.pkcs1',
-                    payload_type: 'com.apple.security.pkcs1',
-                    payload_identifier: `com.addigy.securityAndPrivacy.com.apple.security.pkcs1.${groupUUID}`,
-                    payload_uuid: uuidv4(),
-                    is_from_security_profile: true,
-                },
-                {
-                    ...basePayload,
-                    addigy_payload_type:
-                        'com.addigy.securityAndPrivacy.com.apple.security.FDERecoveryKeyEscrow',
-                    payload_type: 'com.apple.security.FDERecoveryKeyEscrow',
-                    payload_identifier: `com.addigy.securityAndPrivacy.com.apple.security.FDERecoveryKeyEscrow.${groupUUID}`,
-                    payload_uuid: uuidv4(),
-                    encrypt_cert_payload_uuid: encryptCertPayloadUUID,
-                    location: 'Key will be escrowed to an Addigy secure database.',
-                },
-                {
-                    ...basePayload,
-                    addigy_payload_type:
-                        'com.addigy.securityAndPrivacy.com.apple.security.FDERecoveryRedirect',
-                    payload_type: 'com.apple.security.FDERecoveryRedirect',
-                    payload_identifier: `com.addigy.securityAndPrivacy.com.apple.security.FDERecoveryRedirect.${groupUUID}`,
-                    payload_uuid: uuidv4(),
-                    encrypt_cert_payload_uuid: encryptCertPayloadUUID,
-                    redirect_url: '',
-                },
-            )
-
-        try {
-            let res = await this._addigyRequest(
-                'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'POST',
-                    json: { payloads },
-                },
-            )
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    /**
-     *
-     * @param name
-     * name of the Policy
-     * @param pppcPolicy
-     * Enter in the identifier, code_requirement and an array of service items the identifier and code requirement will be assigned to
-     * example: {identifier: 'ex_identifier', codeRequirement: 'ex_coderequirement', services: [{service: 'address_book', identifier_type: "bundleID", allowed: true}, {service: "screen_capture", identifier_type: "bundleID", authorization: "AllowStandardUserToSetSystemService"}] }
-     */
-
-    async createPPPCPolicy(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        pppcPolicy: PPPCInput[],
-    ): Promise<any> {
-        const groupUUID = uuidv4()
-        const payload: PPPCPayload = {
-            addigy_payload_type:
-                'com.addigy.TCC.configuration-profile-policy.com.apple.TCC.configuration-profile-policy',
-            payload_type: 'com.apple.TCC.configuration-profile-policy',
-            payload_display_name: name,
-            payload_group_id: groupUUID,
-            payload_version: 1,
-            payload_identifier: `com.addigy.TCC.configuration-profile-policy.com.apple.TCC.configuration-profile-policy.${groupUUID}`,
-            payload_uuid: uuidv4(),
-            services: {
-                accessibility: [],
-                address_book: [],
-                apple_events: [],
-                calendar: [],
-                camera: [],
-                microphone: [],
-                photos: [],
-                post_event: [],
-                reminders: [],
-                system_policy_all_files: [],
-                system_policy_sys_admin_files: [],
-                file_provider_presence: [],
-                listen_event: [],
-                media_library: [],
-                screen_capture: [],
-                speech_recognition: [],
-                system_policy_desktop_folder: [],
-                system_policy_documents_folder: [],
-                system_policy_downloads_folder: [],
-                system_policy_network_volumes: [],
-                system_policy_removable_volumes: [],
-            },
-        }
-        pppcPolicy.forEach((pppc) => {
-            pppc.services.forEach((e) => {
-                const service: PPPCService = {
-                    allowed: false,
-                    authorization: '',
-                    code_requirement: pppc.codeRequirement,
-                    comment: '',
-                    identifier_type: e.identifierType,
-                    identifier: pppc.identifier,
-                    static_code: e.staticCode ?? false,
-                    predefined_app: null,
-                    manual_selection: true,
-                    rowId: uuidv4(),
-                }
-                if (e.service === 'screen_capture' && e.authorization) {
-                    service.authorization = e.authorization
-                }
-                if (e.service !== 'screen_capture') service.allowed = e.allowed
-
-                if (e.service === 'apple_events') {
-                    service.ae_receiver_identifier = e.aeReceiverIdentifier
-                    service.ae_receiver_identifier_type = e.aeReceiverIdentifierType
-                    service.ae_receiver_code_requirement = e.aeReceiverCodeRequirement
-                    service.ae_receiver_predefined_app = null
-                    service.ae_receiver_manual_selection = true
-                }
-                payload.services[e.service].push(service)
-            })
-        })
-
-        const res = await this._addigyRequest(
-            'https://app-prod.addigy.com/api/mdm/user/profiles/configurations',
-            {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                },
-                method: 'POST',
-                json: { payloads: [payload] },
-            },
-        )
-        return JSON.parse(res.body)
-    }
-
-    async createCustomFact(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-        script: string,
-        scriptType: 'bash' | 'python' | 'zsh',
-    ): Promise<CustomFact> {
-        const shebang = {
-            bash: '#!/bin/bash',
-            python: '#!/usr/bin/python',
-            zsh: '#!/bin/zsh',
-        }
-        const body = {
-            name,
-            os_architectures: {
-                linux_arm: {
-                    is_supported: false,
-                    language: '',
-                    shebang: '',
-                    script: '',
-                },
-                darwin_amd64: {
-                    is_supported: true,
-                    language: scriptType,
-                    shebang: shebang[scriptType],
-                    script,
-                },
-            },
-            return_type: 'string',
-        }
-        const res = await this._addigyRequest(
-            'https://app-prod.addigy.com/api/services/facts/custom',
-            {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                },
-                method: 'POST',
-                json: body,
-            },
-        )
-        return JSON.parse(res.body)
-    }
-
-    async getCustomFacts(authObject: IAddigyInternalAuthObject): Promise<CustomFact[]> {
-        const res = await this._addigyRequest(
-            'https://app-prod.addigy.com/api/services/facts/custom',
-            {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                },
-                method: 'GET',
-            },
-        )
-
-        const results = JSON.parse(res.body)
-        return results.custom_facts ?? []
-    }
-
-    async getCustomFactByName(
-        authObject: IAddigyInternalAuthObject,
-        name: string,
-    ): Promise<CustomFact | undefined> {
-        const facts = await this.getCustomFacts(authObject)
-        return facts.find((e: any) => e.name === name)
-    }
-
-    async getMdmConfigurations(
-        authObject: IAddigyInternalAuthObject,
-    ): Promise<MdmConfigurationPayload[]> {
-        try {
-            let res = await this._addigyRequest(
-                'https://app.addigy.com/api/v2/mdm/configurations/profiles',
-                {
-                    headers: {
-                        Cookie: `auth_token=${authObject.authToken};`,
-                        origin: 'https://app-prod.addigy.com',
-                    },
-                    method: 'GET',
-                },
-            )
-            return JSON.parse(res.body)?.payloads
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getMdmConfigurationByName(authObject: IAddigyInternalAuthObject, name: string) {
-        try {
-            const mdmConfigurations = await this.getMdmConfigurations(authObject)
-            return mdmConfigurations.find((e) => e.payload_display_name === name)
-        } catch (err) {
-            throw err
-        }
-    }
-    async getFileVaultKeys(authObject: IAddigyInternalAuthObject): Promise<object[]> {
-        try {
-            let res = await this._addigyRequest('https://prod.addigy.com/get_org_filevault_keys/', {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                },
-                method: 'GET',
-            })
-            return JSON.parse(res.body)
-        } catch (err) {
-            throw err
-        }
-    }
-
-    toSnakeCase(text: string) {
-        // Complex and complete regex string thanks to https://github.com/zellwk/javascript/issues/14
-        return text
-            .replace(
-                /([^\p{L}\d]+|(?<=\p{L})(?=\d)|(?<=\d)(?=\p{L})|(?<=[\p{Ll}\d])(?=\p{Lu})|(?<=\p{Lu})(?=\p{Lu}\p{Ll})|(?<=[\p{L}\d])(?=\p{Lu}\p{Ll}))/gu,
-                '_',
-            )
-            .toLowerCase()
-    }
-
-    async getApnsCerts(
-        authObject: IAddigyInternalAuthObject,
-        next?: string,
-        previous?: string,
-    ): Promise<object[]> {
-        let url = 'https://app-prod.addigy.com/api/apn/user/apn/list'
-        if (next) {
-            url = `${url}?next=${next}`
-        }
-        if (previous) {
-            url = `${url}?previous=${previous}`
-        }
-
-        try {
-            let res = await this._addigyRequest(url, {
-                headers: {
-                    Cookie: `auth_token=${authObject.authToken};`,
-                    origin: 'https://app-prod.addigy.com',
-                    email: authObject.emailAddress,
-                    orgid: authObject.orgId,
-                },
-                method: 'GET',
-            })
-            return JSON.parse(res.body).items
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getAuthObject(): Promise<IAddigyInternalAuthObject> {
-        let postBody: any = {
-            username: this.config.adminUsername,
-            password: this.config.adminPassword,
-        }
-
-        try {
-            if (!this.config.adminUsername || !this.config.adminPassword)
-                throw new Error(
-                    "The function you are using hits Addigy's internal API, but no username or password was provided in the constructor. Please fill out the adminUsername and adminPassword parameters.",
-                )
-            let res = await this._addigyRequest('https://prod.addigy.com/signin/', {
-                method: 'POST',
-                json: postBody,
-            })
-
-            let authObject = {
-                orgId: res.body.orgid,
-                authToken: res.body.authtoken,
-                emailAddress: res.body.email,
-            }
-
-            return authObject
-        } catch (err) {
-            throw err
-        }
-    }
-
-    async getImpersonationAuthObject(
-        authObject: IAddigyInternalAuthObject,
-        orgId: string,
-    ): Promise<IAddigyInternalAuthObject> {
-        let postBody: any = {
-            parent_orgid: authObject.orgId,
-            child_orgid: orgId,
-            user_email: authObject.emailAddress,
-        }
-
-        try {
-            let res = await this._addigyRequest('https://app.addigy.com/api/impersonation', {
-                headers: {
-                    Cookie: `prod_auth_token=${authObject.authToken};`,
-                    origin: 'https://app.addigy.com',
-                },
-                method: 'POST',
-                json: postBody,
-            })
-
-            let impersonationAuthObject = {
-                orgId: orgId,
-                authToken: res.headers['set-cookie']
-                    .find(
-                        (e: string) =>
-                            e.includes('prod_auth_token') && !e.includes('original_auth_token'),
-                    )
-                    .split('auth_token=')[1]
-                    .split(';')[0],
-                emailAddress: authObject.emailAddress,
-            }
-
-            return impersonationAuthObject
-        } catch (err) {
-            throw err
-        }
-    }
-
-    private async _addigyRequest(url: string, options: any): Promise<any> {
-        try {
-            let res = await got(url, options)
-            return res
-        } catch (err) {
-            throw err
-        }
-    }
-}
+export { Addigy } from './lib/addigy'
+export { Alerts } from './lib/alerts/alerts'
+export * from './lib/alerts/alert.types'
+export { Applications } from './lib/apps/applications'
+export { Billing } from './lib/billing/billing'
+export { Certs } from './lib/certs/certs'
+export { Commands } from './lib/commands/commands'
+export { Configs } from './lib/configs/configs'
+export { Devices } from './lib/devices/devices'
+export { Facts } from './lib/facts/facts'
+export { FileVault } from './lib/file-vault/file-vault'
+export { Files } from '././lib/files/files'
+export { Integrations } from './lib/integrations/integrations'
+export { Maintenance } from './lib/maintenance/maintenance'
+export { Policies } from './lib/policies/policies'
+export { ScreenConnect } from './lib/screenconnect/screenconnect'
+export { Software } from './lib/software/software'
+export { Users } from './lib/users/users'
+export * from './lib/users/user.types'
