+++ src/lib/alerts/alert.types.ts
@@ -1,0 +1,5 @@
+export enum AlertStatus {
+    Acknowledged = 'Acknowledged',
+    Resolved = 'Resolved',
+    Unattended = 'Unattended',
+}
