+++ src/lib/alerts/alerts.spec.ts
@@ -1,0 +1,3 @@
+describe('Alerts', () => {
+    it('works', () => expect(true).toBe(true))
+})
