+++ src/lib/apps/applications.spec.ts
@@ -1,0 +1,3 @@
+describe('Applications', () => {
+    it('works', () => expect(true).toBe(true))
+})
