+++ src/lib/billing/billing.spec.ts
@@ -1,0 +1,3 @@
+describe('Billing', () => {
+    it('works', () => expect(true).toBe(true))
+})
