+++ src/lib/commands/commands.spec.ts
@@ -1,0 +1,3 @@
+describe('Commands', () => {
+    it('works', () => expect(true).toBe(true))
+})
