+++ src/lib/configs/configs.spec.ts
@@ -1,0 +1,3 @@
+describe('Configs', () => {
+    it('works', () => expect(true).toBe(true))
+})
