+++ src/lib/devices/devices.spec.ts
@@ -1,0 +1,3 @@
+describe('Devices', () => {
+    it('works', () => expect(true).toBe(true))
+})
