+++ src/lib/facts/facts.spec.ts
@@ -1,0 +1,3 @@
+describe('Facts', () => {
+    it('works', () => expect(true).toBe(true))
+})
