+++ src/lib/file-vault/file-vault.spec.ts
@@ -1,0 +1,3 @@
+describe('FileVault', () => {
+    it('works', () => expect(true).toBe(true))
+})
