+++ src/lib/files/files.spec.ts
@@ -1,0 +1,3 @@
+describe('Files', () => {
+    it('works', () => expect(true).toBe(true))
+})
