+++ src/lib/integrations/integrations.spec.ts
@@ -1,0 +1,3 @@
+describe('Integrations', () => {
+    it('works', () => expect(true).toBe(true))
+})
