+++ src/lib/maintenance/maintenance.spec.ts
@@ -1,0 +1,3 @@
+describe('Maintenance', () => {
+    it('works', () => expect(true).toBe(true))
+})
