+++ src/lib/policies/policies.spec.ts
@@ -1,0 +1,3 @@
+describe('Policies', () => {
+    it('works', () => expect(true).toBe(true))
+})
