+++ src/lib/profiles/profiles.spec.ts
@@ -1,0 +1,3 @@
+describe('Profiles', () => {
+    it('works', () => expect(true).toBe(true))
+})
