+++ src/lib/screenconnect/screenconnect.spec.ts
@@ -1,0 +1,3 @@
+describe('ScreenConnect', () => {
+    it('works', () => expect(true).toBe(true))
+})
