+++ src/lib/software/software.spec.ts
@@ -1,0 +1,3 @@
+describe('Software', () => {
+    it('works', () => expect(true).toBe(true))
+})
