+++ src/lib/users/user.types.ts
@@ -1,0 +1,5 @@
+export enum UserRoles {
+    Owner = 'power',
+    Admin = 'admin',
+    User = 'user',
+}
