+++ src/util/index.ts
@@ -1,0 +1,1 @@
+export { toSnakeCase } from './toSnakeCase'
